Moved utilities functions to their place
